requirements.txt
pandas
requests==2.23.0
glom
discord
paho-mqtt
